Enhanced the filtering, automation, update mode.
Enhanced the rules by which items are filtered
and sorted. Also added the ability to update an
existing release draft so that I no longer need
to delete and recreate. This is particularly
useful for updating via CI. Also adds two shell
scripts, one to create a release, and another
to update a release.